+++ components/nfceditor/impl/src/main/java/com/flipperdevices/nfceditor/impl/api/NfcEditorApiImpl.kt
@@ -3,6 +3,8 @@ package com.flipperdevices.nfceditor.impl.api
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.keyparser.api.model.FlipperKeyParsed
 import com.flipperdevices.metric.api.MetricApi
+import com.flipperdevices.metric.api.events.complex.DebugInfoEnum
+import com.flipperdevices.metric.api.events.complex.DebugInfoEvent
 import com.flipperdevices.nfceditor.api.NfcEditorApi
 import com.squareup.anvil.annotations.ContributesBinding
 import javax.inject.Inject
@@ -20,5 +22,26 @@ class NfcEditorApiImpl @Inject constructor(
         return parsedKey is FlipperKeyParsed.NFC &&
             SUPPORTED_NFC_FORMATS.contains(parsedKey.version) &&
             parsedKey.deviceType == SUPPORTED_NFC_TYPE
+    }
+
+    override fun reportUnsupportedFormat(parsedKey: FlipperKeyParsed) {
+        if (parsedKey !is FlipperKeyParsed.NFC) {
+            // Report only NFC
+            return
+        }
+        if (parsedKey.deviceType != SUPPORTED_NFC_TYPE) {
+            // Report only Mifare Classic formats
+            return
+        }
+        if (SUPPORTED_NFC_FORMATS.contains(parsedKey.version)) {
+            // Report only unsupported format
+            return
+        }
+        metricApi.reportComplexEvent(
+            DebugInfoEvent(
+                DebugInfoEnum.NFC_UNSUPPORTED_EDIT,
+                parsedKey.version.toString()
+            )
+        )
     }
 }